refactor: Вынесла Navbar в share
