+++ packages/client/src/shared/ui/index.ts
@@ -1,2 +1,3 @@
 export { CounterAnimation } from './CounterAnimation'
 export { FileInput } from './FileInput'
+export { Navbar } from './Navbar'
