+++ packages/client/src/widgets/Header/model/navbar-items.ts
@@ -1,0 +1,6 @@
+export const navbarItems = [
+  { key: '/', label: 'Главная' },
+  { key: '/game', label: 'Играть' },
+  { key: '/leaderboard', label: 'Лидеры' },
+  { key: '/forum', label: 'Форум' },
+]
